Update 0.01D
Created Main Menu Scene
Added a basic home screeen which only consists of a small platform for the player to stand on and a UI canvas
Created an extra ray interactor that is completely transparent except when hovering over UI elements. Interactor only works with UI
Player can press "Play" button to move to next scene and "Exit" button to close game

Created a basic system to move between scenes. "Game" scene, which is currently just the sample scene, is loaded asynchronously when in the main menu scene. Player is transported when the press "Play"

Added post processing and a glowing material. No actual reason for this one. I just thought it looked cool