Update 0.01G
Now would be a good time to mention that I'm prolly not going all the way to Z. The 0.01 section of this is until I complete the basics of the game. As it stands, I need to have the gun fire properly, targets appear and disappear correctly, create a scene which will use the sonar mechanic, find a way to change the player hand model with a script, and give the player a physical body. Once all that is done, We will be moving onto 0.10A, where I will begin the implementation of multiplayer.

Now onto this update
--------------------------
Updated HitTarget Script so that targets correctly track the player's position
Targets also now properly move from active to inactive lists when shot at

Renamed Gun Controller to BulletController