0.01F1
Little mini changes just because

BulletController and Gun have been separated into 2 different scripts,. Bullet Controller is an empty gameobject that creates and tracks all the bullets in the scene. The gun has a reference to the bullet controller so that when Fire() is called, the bullet controller will just have bullets put in place

Created the script that will end up changing the player hand model from hands to guns and back.