Update 0.01H
I seemed to forget the alphabet

Big changes this time, though the number is mostly from me trying to get a mirror to work

Added a model for the player's head and attached it to the player prefab.

Created a new type of target that moves back and forth in the x/y/z direction

Made the hitTarget script a base for all other scripts. Oscillating and Strafing targets extend from HitTarget.cs

Updated target controller to spawn different target types

Made a mirror
---------------

Next update, we beging the multiplayer shenanigans(probably)